checkQUoteDatabase() with quoteLineNum updated

+++ CuriApp/View/Main Flow/Library/Book/LinesView.swift
@@ -42,8 +42,9 @@ struct LinesView: View {
     
     var body: some View {
         VStack (spacing: curiSpacing(.sp8)) {
-            ForEach(bookLinesOriginal, id: \.self) { line in
-                let quote = Quote(bookID: bookID, quoteBook: bookTitle, quoteAuthor: bookAuthor, quoteContent: line, quoteHighlight: bookViewModel.selectedPen ?? pencilDatabase[0], isConnected: false)
+            ForEach(Array(bookLinesOriginal.enumerated()), id: \.offset) { lineNum, line in
+                
+                let quote = Quote(bookID: bookID, quoteID: UUID(), quoteLineNum: lineNum + 1, quoteBook: bookTitle, quoteAuthor: bookAuthor, quoteContent: line, quoteHighlight: bookViewModel.selectedPen ?? pencilDatabase[0], isConnected: false)
                 
                 Text(quote.quoteContent)
                     .curiTypo(.bkRegular16)
@@ -51,13 +52,14 @@ struct LinesView: View {
                     .onLongPressGesture(minimumDuration: 0.1) {
                         print("- Line: \(quote.quoteContent)")
                         print("- ID: \(quote.quoteID))")
+                        print("- Line Num: \(quote.quoteLineNum)")
                         
                         checkQuoteDatabase(checkingQuote: quote)
                         
                     }
                     .onTapGesture {
                         withAnimation {
-                            if let existingQuote = quoteDatabase.first(where: { $0.quoteContent == line }) {
+                            if let existingQuote = quoteDatabase.first(where: { $0.quoteContent == quote.quoteContent && $0.quoteLineNum == quote.quoteLineNum }) {
                                 if bookViewModel.selectedLine?.quoteID == existingQuote.quoteID {
                                     bookViewModel.selectedLine = nil
                                 } else {
@@ -69,12 +71,12 @@ struct LinesView: View {
                             bookViewModel.pageIsSelected = true
                         }
                         print("-- Ready to add Note: \(bookViewModel.selectedLine != nil ? "✅" : "🙅🏻‍♂️") - \(String(describing: bookViewModel.selectedLine?.quoteContent)), \(String(describing: bookViewModel.selectedLine?.quoteHighlight.name))")
-                        print("\(quote.quoteID)) - \(quote.quoteContent)")
+                        print("\(quote.quoteLineNum)) - \(quote.quoteContent)")
                     }
                     .foregroundStyle(
                         {
-                            if let quote = quoteDatabase.first(where: { $0.quoteContent == line }) {
-                                if bookViewModel.selectedLine?.quoteContent == line {
+                            if let quote = quoteDatabase.first(where: { $0.quoteContent == quote.quoteContent && $0.quoteLineNum == quote.quoteLineNum }) {
+                                if bookViewModel.selectedLine?.quoteContent == quote.quoteContent {
                                     return Color(quote.quoteHighlight.highlightedTextColor)
                                 } else {
                                     return Color(quote.quoteHighlight.highlightedTextColor)
@@ -86,8 +88,8 @@ struct LinesView: View {
                     )
                     .background(
                         {
-                            if let quote = quoteDatabase.first(where: { $0.quoteContent == line }) {
-                                if bookViewModel.selectedLine?.quoteContent == line {
+                            if let quote = quoteDatabase.first(where: { $0.quoteContent == quote.quoteContent && $0.quoteLineNum == quote.quoteLineNum }) {
+                                if bookViewModel.selectedLine?.quoteContent == quote.quoteContent {
                                     return Color(quote.quoteHighlight.selectedHighlightedBackgroundColor)
                                 } else {
                                     return Color(quote.quoteHighlight.defaultHighlightedBackgroundColor)
@@ -109,7 +111,7 @@ struct LinesView: View {
     
     func checkQuoteDatabase(checkingQuote: Quote) {
         //        if let existingQuote = quoteDatabase.first(where: { $0.quoteID == checkingQuote.quoteID }) {
-        if let existingQuote = quoteDatabase.first(where: { $0.quoteContent == checkingQuote.quoteContent }) {
+        if let existingQuote = quoteDatabase.first(where: { $0.quoteContent == checkingQuote.quoteContent && $0.quoteLineNum == checkingQuote.quoteLineNum }) {
             HapticsManager.access.play(haptics: .light, vibrationEnabledInApp: bookViewModel.vibrationInApp)
             SoundManager.access.play(sound: .highlightRemoved, soundEnabledInApp: bookViewModel.soundInApp)
             modelContext.delete(existingQuote)
@@ -118,8 +120,9 @@ struct LinesView: View {
             SoundManager.access.play(sound: .highlightAdded, soundEnabledInApp: bookViewModel.soundInApp)
             modelContext.insert(checkingQuote)
         }
+        
         WidgetDataManager().updateQuoteOnWidget(quoteDatabase: quoteDatabase)
-//        WidgetCenter.shared.reloadAllTimelines()
+        
     }
